requirements.txt
gunicorn
flask
python-dotenv
google-genai
pdfplumber
python-docx
Pillow
pymupdf
